Refactor UDS.DTC to UDS.OBD2.DTC to make room for another DTC struct
Diagnostic Trouble Codes seem to come in a multitude of formats, the most common one being
used for OBD2 (defined in SAE2012) which maps two bytes into a 5 letter code (e.g., P0170).
UDS has support for codes with a length up to three + 1 bytes and a variety of formats, hence
it makes more sense to have a dedicated model object here.